Export of internal ZetaSQL changes.
--
Change by John Fremlin <fremlin@google.com>:
Tests for setting variables with more complex names
--
Change by ZetaSQL Team <no-reply@google.com>:
Add some parser tests for untested corner cases of the EXCEPT keyword.
--
Change by Scott Wegner <swegner@google.com>:
Small clarification to EvaluatorTableIterator example usage.
Without a `break` statement, execution will fall through to the `Do something with 'iter->GetValue(...)'` block. But since `iter-NextRow()` returned false, there are no values left, and execution should break out of the `while` loop.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add some tests for query parameters and system variable corner cases, including several broken cases for sysvars.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add a parser query reproing an issue with the EXCEPT keyword and macros.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add an overload for FunctionArgumentTypeOptions::Lambda to allow naming of lambda arguments.
--
Change by ZetaSQL Team <no-reply@google.com>:
Adding formatting changes to output inlined in code samples
--
Change by ZetaSQL Team <no-reply@google.com>:
Return a deprecation warning when bare offsetting into arrays of protocol buffer map entries.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove redundant lhs TableScan column_list populated in ArrayScan
--
Change by Matthew Brown <matthewbr@google.com>:
Add --strict_name_resolution_mode flag
--
Change by ZetaSQL Team <no-reply@google.com>:
Create StringToDateCaster.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ST_DUMPPOINTS to docs.
--
Change by ZetaSQL Team <no-reply@google.com>:
Do not generate a ResolvedAST chained builder setter for vector fields, since we do not have vectorized builders.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add test cases for PIVOT with collation. We need to fix reference implementation as described in b/270061671.
--
Change by ZetaSQL Team <no-reply@google.com>:
Flip in_development for FEATURE_V_1_4_SINGLE_TABLE_NAME_ARRAY_PATH
--
Change by ZetaSQL Team <no-reply@google.com>:
Resolve bison warnings about rule type mismatch. In these cases the relevant non-termininal rules do not return nodes, but were declared to do so.
--
Change by ZetaSQL Team <no-reply@google.com>:
In the DDL, update notes for features that are supported, but undocumented.
--
Change by John Fremlin <fremlin@google.com>:
Fix test for efficient compiles using no stack
--
Change by ZetaSQL Team <no-reply@google.com>:
special case for offsetting into arrays of proto map entries
--
Change by ZetaSQL Team <no-reply@google.com>:
Support RANGE types in the resolver for the CREATE TABLE DDL statement.
--
Change by ZetaSQL Team <no-reply@google.com>:
INTERNAL
--
Change by ZetaSQL Team <no-reply@google.com>:
BugFix: Special characters in JSON key are not escaped during write causing incorrect results to be returned.
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix unparsing of SELECT (0).a
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix the issue in SQLBuilder that caused 'AS VALUE' missed from the output SQL when the direct child of an EXPORT DATA statement is a SET operation whose children are SELECT AS VALUE.
--
Change by Honglin Zheng <hlzheng@google.com>:
Introduce CREATE_MODEL_WITH_ALIASED_QUERY language feature flag and implement error checking.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add ParseLocationRange to ResolvedExpressionColumn.
--
Change by ZetaSQL Team <no-reply@google.com>:
Complex nested UNNEST() testcase
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_features2" compliance test option.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in float_and_double_queries.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in groupby_queries.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in cast_function.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in dml_update.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in with_recursive.test.
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in array_functions.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in aggregation_order_by_queries.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in concat_function.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in safe_function.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Remove "test_featuresN=" use in civil_time.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Update some compliance test files to be consistent with the output of update_goldens.py
--
Change by Honglin Zheng <hlzheng@google.com>:
Rename ASTWithClauseEntry to ASTAliasedQuery.
--
Change by ZetaSQL Team <no-reply@google.com>:
Use RET_CHECK() instead of DCHECK() in sql_builder code to avoid crashing on failure.
--
Change by ZetaSQL Team <no-reply@google.com>:
Partially migrates BuiltinFunctionInliner to ResolvedASTRewriteVisitor to reduce copying. I say partially because it still uses AnalyzeSubstitute, which is still doing full deep copies of the function arguments (see b/269806964).
--
Change by ZetaSQL Team <no-reply@google.com>:
Add a check on the truncated timestamp to avoid timestamp underflow for function TIMESTAMP_TRUNC.
--
Change by ZetaSQL Team <no-reply@google.com>:
Partially migrates BuiltinFunctionInliner to ResolvedASTRewriteVisitor to reduce copying. I say partially because it still uses AnalyzeSubstitute, which is still doing full deep copies of the function arguments (see b/269806964).
--
Change by ZetaSQL Team <no-reply@google.com>:
Add WITH DIFFERENTIAL_PRIVACY compliance tests that use privacy_unit_column.
--
Change by ZetaSQL Team <no-reply@google.com>:
Print the parse location in standalone_expr.test
--
Change by ZetaSQL Team <no-reply@google.com>:
Tweak grammar
--
Change by ZetaSQL Team <no-reply@google.com>:
Fix wording around inserting into columns with default values.
--
Change by ZetaSQL Team <no-reply@google.com>:
Add missing compliance tests for the previously added logic of replacing k_threshold_expr for unique users count column (with report PROTO).
--
Change by ZetaSQL Team <no-reply@google.com>:
Correcting a `FILTER_FIELDS` example which incorrectly shows the filtered out field name instead of the remaining field name (though the values are correct).
  (And 330 more changes)

GitOrigin-RevId: 8766037625dc0964d5e0ea9e08bebf4e3e2e86f2

+++ zetasql/public/civil_time.cc
@@ -222,9 +222,12 @@ TimeValue TimeValue::FromPacked32SecondsAndMicros(
 }
 
 int32_t TimeValue::Packed32TimeSeconds() const {
-  return (hour_ << kHourShift) |
-         (minute_ << kMinuteShift) |
-         (second_ << kSecondShift);
+  // When the Hour-Minute-Second value is incorrect, hour_ is set to -1.
+  // It is undefied behavior to left shift a negative number, hence hour_
+  // is static_cast-ed to uin8t before being left-shifted and packed into
+  // a int32_t value. (b/262060332)
+  return (static_cast<uint8_t>(hour_) << kHourShift) |
+         (minute_ << kMinuteShift) | (second_ << kSecondShift);
 }
 
 int64_t TimeValue::Packed64TimeMicros() const {
